requirements.txt

python-dotenv==1.0.1

requests
Flask>=3.0
gunicorn>=21.2
openai>=1.40
twilio>=9.0
